+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
+- `GHPageServer` class `start` method modified
 - Dependencies structure modified
 - Test system modified
 ## [0.4] - 2026-05-08